feat: consume rating API

+++ app/src/main/java/com/c22ho01/hotelranking/data/remote/response/hotel/HotelData.kt
@@ -25,7 +25,7 @@ data class HotelData(
     val neighborhood: String,
 
     @SerializedName("hotel_star")
-    val star: Double,
+    val star: Float,
 
     @SerializedName("price_per_night")
     val pricePerNight: Int,
